simple split of DestinationAcceptanceTest
